fix(build): use .python-version to specify python version
This replaces the runtime.txt file with a .python-version file,
which is a different method for declaring the Python version that
Render.com supports.

This is intended to work around a platform-level issue where Render
is ignoring the runtime.txt file and defaulting to Python 3.13, even
after a cache clear.